TIP-860: Move Channel related api tests
