+++ tests/back/Channel/EndToEnd/Channel/ExternalApi/ListChannelEndToEnd.php
@@ -1,12 +1,11 @@
 <?php
 
-namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller\Channel;
+namespace AkeneoTest\Channel\EndToEnd\Channel\ExternalApi;
 
-use Akeneo\Test\Integration\Configuration;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-class ListChannelIntegration extends ApiTestCase
+class ListChannelEndToEnd extends ApiTestCase
 {
     public function testListChannels()
     {
